+++ project/src/components/app/app.tsx
@@ -5,13 +5,14 @@ import AddReview from '../../pages/add-review';
 import SignIn from '../../pages/sign-in';
 import Player from '../../pages/player';
 import MyList from '../../pages/my-list';
-import MoviePage from '../../pages/movie-page';
 import PrivateRoute from '../private-route/private-route';
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import { useAppSelector } from '../../hooks';
 import { LoadingScreen } from '../loading-screen/loading-screen';
 import { getIsLoadingFilms } from '../../store/data-api-process/selectors';
 import { getAuthorizationStatus } from '../../store/user-process/selectors';
+import { MoviePage } from '../../pages/movie-page';
+
 
 function App(): JSX.Element {
   const isLoadingFilms = useAppSelector(getIsLoadingFilms);
