+++ project/src/components/main-components/main-genre-menu.tsx
@@ -1,8 +1,8 @@
 import { MouseEvent } from 'react';
 import { FilterMainNavMenu, GenresFilter } from '../../const';
 import { useAppDispatch, useAppSelector } from '../../hooks';
-import { mainFilterChangAction } from '../../store/main-genre-filter-process/main-genre-filter-process';
-import { getFilter } from '../../store/main-genre-filter-process/selectors';
+import { mainFilterChangAction } from '../../store/filter-process/filter-process';
+import { getFilter } from '../../store/filter-process/selectors';
 
 
 function MainGenreMenu ():JSX.Element {
