+++ project/src/components/main-components/main-show-more-button.tsx
@@ -1,5 +1,5 @@
 import { useAppSelector } from '../../hooks';
-import { getFilmFilterCount } from '../../store/main-genre-filter-process/selectors';
+import { getFilmFilterCount } from '../../store/filter-process/selectors';
 
 
 const SLICE_STEP = 8;
