+++ project/src/pages/movie-page.tsx
@@ -1,26 +1,20 @@
-import { Link, Navigate, useParams } from 'react-router-dom';
-import { useState, MouseEvent, useEffect } from 'react';
+import { Navigate, useParams } from 'react-router-dom';
+import {memo, useEffect} from 'react';
 import ArtBoard from '../components/art-board/art-board';
-import { AppRoute, AuthorizationStatus, CatalogFilm, NavMenuMoviePage } from '../const';
-import { AllTabs } from '../components/all-tabs/all-tabs';
+import { AppRoute, CatalogFilm} from '../const';
 import Logo from '../components/logo/logo';
-import { UserSign } from '../components/user-sign/user-sign';
 import { useAppDispatch, useAppSelector } from '../hooks';
 import { getDataOpenFilmAction, getDataReviewsOpenFilm, getDataSimilarFilmsAction } from '../store/api-action';
 import { LoadingScreen } from '../components/loading-screen/loading-screen';
 import { CatalogFilmCardsInterface } from '../components/catalog-film-card/catalog-film-cards-interface';
-import { getAllFilms, getIsErrorResponse, getOpenedFilms, getReviews } from '../store/data-api-process/selectors';
-import { getAuthorizationStatus } from '../store/user-process/selectors';
-import { MyListLink } from '../components/my-list-link/my-list-link';
+import { getIsErrorResponse, getOpenedFilms} from '../store/data-api-process/selectors';
 
-const MoviePage = ():JSX.Element=>{
+import { MovieInfo } from '../components/movie-info/movie-info';
+
+function MoviePageComponents ():JSX.Element {
   const dispatch = useAppDispatch();
-  const [navMenuButtonCount, setNavMenuButtonCount] = useState(NavMenuMoviePage.OVERVIEW);
   const openedFilm = useAppSelector(getOpenedFilms);
-  const allFilms = useAppSelector(getAllFilms);
-  const reviews = useAppSelector(getReviews);
   const isErrorResponse = useAppSelector(getIsErrorResponse);
-  const authorizationStatus = useAppSelector(getAuthorizationStatus);
   const {id:idParam} = useParams();
 
   useEffect(() => {
@@ -39,99 +33,10 @@ const MoviePage = ():JSX.Element=>{
     return <LoadingScreen/>;
   }
 
-  const {
-    id,
-    backgroundColor,
-    bigPoster,
-    filmName,
-    genre,
-    yearCreation,
-    poster,
-    isFavorite
-  } = openedFilm;
-
-  const myListFilmCount = allFilms.filter((filmCard)=>filmCard.isFavorite === true).length;
-
   return (
     <>
       <ArtBoard/>
-      <section className="film-card film-card--full" style={{background:backgroundColor}}>
-        <div className="film-card__hero">
-          <div className="film-card__bg">
-            <img src={bigPoster} alt= {filmName} />
-          </div>
-          <h1 className="visually-hidden">WTW</h1>
-          <header className="page-header film-card__head">
-            <Logo />
-            <UserSign/>
-          </header>
-          <div className="film-card__wrap">
-            <div className="film-card__desc">
-              <h2 className="film-card__title">{filmName}</h2>
-              <p className="film-card__meta">
-                <span className="film-card__genre">{genre}</span>
-                <span className="film-card__year">{yearCreation}</span>
-              </p>
-              <div className="film-card__buttons">
-                <Link to={AppRoute.Player.replace(':id', `${id}`)} className="btn btn--play film-card__button" type="button">
-                  <svg viewBox="0 0 19 19" width="19" height="19">
-                    <use xlinkHref="#play-s"></use>
-                  </svg>
-                  <span>Play</span>
-                </Link>
-                <MyListLink/>
-                {
-                  authorizationStatus === AuthorizationStatus.Auth
-                    ? <Link to={AppRoute.AddReview.replace(':id', `${id}`)} className="btn film-card__button">Add review</Link>
-                    : null
-                }
-              </div>
-            </div>
-          </div>
-        </div>
-        <div className="film-card__wrap film-card__translate-top">
-          <div className="film-card__info">
-            <div className="film-card__poster film-card__poster--big">
-              <img src={poster} alt={filmName} width="218" height="327" />
-            </div>
-            <div className="film-card__desc">
-              <nav className="film-nav film-card__nav">
-                <ul className="film-nav__list">
-                  <li className={`film-nav__item ${navMenuButtonCount === NavMenuMoviePage.OVERVIEW ? 'film-nav__item--active' : ''}`}
-                    onClick={(event:MouseEvent<HTMLElement>) => {
-                      event.preventDefault();
-                      setNavMenuButtonCount(NavMenuMoviePage.OVERVIEW);
-                    }}
-                  >
-                    <a href="#top" className="film-nav__link">Overview</a>
-                  </li>
-                  <li className={`film-nav__item ${navMenuButtonCount === NavMenuMoviePage.DETAILS ? 'film-nav__item--active' : ''}`}
-                    onClick ={(event:MouseEvent<HTMLElement>) => {
-                      event.preventDefault();
-                      setNavMenuButtonCount(NavMenuMoviePage.DETAILS);
-                    }}
-                  >
-                    <a href="#top" className="film-nav__link">Details</a>
-                  </li>
-                  <li className={`film-nav__item ${navMenuButtonCount === NavMenuMoviePage.REVIEWS ? 'film-nav__item--active' : ''}`}
-                    onClick={(event:MouseEvent<HTMLElement>) => {
-                      event.preventDefault();
-                      setNavMenuButtonCount(NavMenuMoviePage.REVIEWS);
-                    }}
-                  >
-                    <a href="#top" className="film-nav__link" >Reviews</a>
-                  </li>
-                </ul>
-              </nav>
-              <AllTabs
-                filmForPage={openedFilm}
-                nameButton={navMenuButtonCount}
-                reviews={reviews}
-              />
-            </div>
-          </div>
-        </div>
-      </section>
+      {<MovieInfo openedFilm={openedFilm}/>}
       <div className="page-content">
         <section className="catalog catalog--like-this">
           <h2 className="catalog__title">More like this</h2>
@@ -148,6 +53,6 @@ const MoviePage = ():JSX.Element=>{
       </div>
     </>
   );
-};
+}
 
-export default MoviePage;
+export const MoviePage = memo(MoviePageComponents);
