+++ project/src/store/root-reducer.ts
@@ -1,13 +1,13 @@
 import { combineReducers} from '@reduxjs/toolkit';
 import { NameSpace } from '../const';
 import {dataAPIProcess} from './data-api-process/data-api-process';
-import { mainGenreFilterProcess } from './main-genre-filter-process/main-genre-filter-process';
+import { filterProcess } from './filter-process/filter-process';
 import { userProcess } from './user-process/user-process';
 
 
 export const rootReducer = combineReducers({
   [NameSpace.DATA]:dataAPIProcess.reducer,
-  [NameSpace.MAIN]:mainGenreFilterProcess.reducer,
+  [NameSpace.FILTER]:filterProcess.reducer,
   [NameSpace.USER]:userProcess.reducer
 });
